refactor: remove feeService.spec.ts and refactor feeService.ts
- Deleted the feeService.spec.ts file, removing all associated tests for AssetAndFeeListMapper.
- Refactored feeService.ts by removing unused imports and the AssetAndFeeListMapper class, simplifying the code structure.
- Updated references to icp-account-id utility to icpAccountId in icpLedger.spec.ts, icpLedger.ts, and walletStore.svelte.ts.
- Renamed link-tx-cart-store.svelte and wallet-tx-cart-store.svelte to linkTxCartStore.svelte and walletTxCartStore.svelte respectively for consistency.
- Updated fee breakdown method calls in LinkTxCart.svelte and WalletTxCart.svelte to use the new buildBreakdown method.
